pr fixes

+++ src/world.h
@@ -38,9 +38,9 @@ typedef struct world_t {
 
 world_t *world_new(uint8_t player_count);
 void world_free(world_t *world);
-void world_generate(world_t *world, unsigned long seed);
-bool world_multiplayer_generate(world_t *world, unsigned long seed);
-void world_generate(world_t *world, unsigned long seed, button_mode_t mode);
+void world_generate(world_t *world, uint16_t seed);
+bool world_multiplayer_generate(world_t *world, uint16_t seed);
+void world_generate(world_t *world, uint16_t seed, button_mode_t mode);
 uint8_t world_count_boxes(world_t *world);
 
 void world_update(world_t *world, uint8_t inputs);
